Components: Replace Text with ui/Text
Text is `experimental` so it's safe to replace it completely with the new text component from `ui`. It's usages in Gutenberg have been updated to use the new API.
