+++ packages/components/src/text/styles.js
