+++ packages/components/src/text/utils.js
