+++ packages/components/src/truncate/README.md
