+++ packages/components/src/truncate/hook.js
