+++ packages/components/src/truncate/index.js
