+++ packages/components/src/truncate/stories/index.js
