+++ packages/components/src/truncate/styles.js
