+++ packages/components/src/truncate/test/truncate.js
