+++ packages/components/src/truncate/types.ts
