+++ packages/components/src/truncate/utils.js
