+++ packages/components/src/ui/index.js
@@ -10,8 +10,6 @@ export * from './h-stack';
 export * from './heading';
 export * from './shortcut';
 export * from './spinner';
-export * from './text';
-export * from './truncate';
 export * from './v-stack';
 export * from './view';
 export * from './visually-hidden';
