+++ packages/edit-site/src/components/header/document-actions/index.js
@@ -91,7 +91,7 @@ export default function DocumentActions( {
 				className="edit-site-document-actions__title-wrapper"
 			>
 				<Text
-					variant="body.small"
+					variant="body"
 					className="edit-site-document-actions__title-prefix"
 				>
 					<VisuallyHidden as="span">
@@ -104,7 +104,7 @@ export default function DocumentActions( {
 				</Text>
 
 				<Text
-					variant="body.small"
+					variant="body"
 					className="edit-site-document-actions__title"
 					as="h1"
 				>
@@ -112,7 +112,7 @@ export default function DocumentActions( {
 				</Text>
 
 				<Text
-					variant="body.small"
+					variant="body"
 					className="edit-site-document-actions__secondary-item"
 				>
 					{ label ?? '' }
